+++ src/yamk/make.py
@@ -4,7 +4,7 @@ import re
 import subprocess
 import sys
 
-import tomlkit
+import tomli
 
 from yamk import lib
 
@@ -30,12 +30,12 @@ class MakeCommand:
                 lambda var: var.split("=", maxsplit=1), args.variables
             )
         ]
-        with open(makefile) as file:
-            parsed_toml = tomlkit.parse(file.read())
+        with open(makefile, "rb") as file:
+            parsed_toml = tomli.load(file)
         makefile_dir = makefile.with_suffix(makefile.suffix + ".d")
         for path in sorted(makefile_dir.glob("*.toml")):
-            with open(path) as file:
-                parsed_toml = lib.deep_merge(parsed_toml, tomlkit.parse(file.read()))
+            with open(path, "rb") as file:
+                parsed_toml = lib.deep_merge(parsed_toml, tomli.load(file))
         self.globals = parsed_toml.pop("$globals", {})
         self._parse_recipes(parsed_toml)
         self.subprocess_kwargs = {